requirements/production.txt
Django==1.11.3
psycopg2==2.7.1
prettyconf==1.2.3
dj-database-url==0.4.2
djangorestframework==3.6.3
django-mptt==0.8.7
djangorestframework-recursive==0.1.2